Make position in CRDTCharacter hold long values instead of int
fix #13

+++ include/CRDTCharacter.hpp
@@ -13,15 +13,15 @@ class CRDTCharacter {
   char value;
   int counter;
   std::string site_id;
-  std::vector<int> positions;
+  std::vector<long> positions;
 
  public:
-  explicit CRDTCharacter(char value, int counter, std::string &site_id, std::vector<int> &positions);
+  explicit CRDTCharacter(char value, int counter, std::string &site_id, std::vector<long> &positions);
 
   char GetValue() const;
   int GetCounter() const;
   const std::string &GetSiteId() const;
-  const std::vector<int> &GetPositions() const;
+  const std::vector<long> &GetPositions() const;
 };
 
 #endif //COLLABORATIVE_TEXT_EDITOR_SRC_CRDT_CRDTCHARACTER_HPP_
